refactor(cursors): split cursor.rs into modules
