+++ src/bytes_ext.rs
@@ -29,7 +29,6 @@ impl BytesExt {
         self.cursor = self.bytes.len() - n;
     }
 
-    #[cfg(any(test, feature = "lz4", feature = "watch"))]
     #[inline(always)]
     pub(crate) fn advance(&mut self, n: usize) {
         debug_assert!(n <= self.remaining());
