+++ src/cursors/json.rs
@@ -1,0 +1,69 @@
+use crate::{
+    bytes_ext::BytesExt,
+    cursors::RawCursor,
+    error::{Error, Result},
+    response::Response,
+};
+use serde::Deserialize;
+use std::marker::PhantomData;
+
+pub(crate) struct JsonCursor<T> {
+    raw: RawCursor,
+    bytes: BytesExt,
+    line: String,
+    _marker: PhantomData<T>,
+}
+
+// We use `JSONEachRowWithProgress` to avoid infinite HTTP connections.
+// See https://github.com/ClickHouse/ClickHouse/issues/22996 for details.
+#[derive(Deserialize)]
+#[serde(rename_all = "lowercase")]
+enum JsonRow<T> {
+    Row(T),
+    Progress {},
+}
+
+impl<T> JsonCursor<T> {
+    const INITIAL_BUFFER_SIZE: usize = 1024;
+
+    pub(crate) fn new(response: Response) -> Self {
+        Self {
+            raw: RawCursor::new(response),
+            bytes: BytesExt::default(),
+            line: String::with_capacity(Self::INITIAL_BUFFER_SIZE),
+            _marker: PhantomData,
+        }
+    }
+
+    pub(crate) async fn next<'a, 'b: 'a>(&'a mut self) -> Result<Option<T>>
+    where
+        T: Deserialize<'b>,
+    {
+        use bytes::Buf;
+        use std::io::BufRead;
+
+        loop {
+            self.line.clear();
+
+            let read = match self.bytes.slice().reader().read_line(&mut self.line) {
+                Ok(read) => read,
+                Err(err) => return Err(Error::Custom(err.to_string())),
+            };
+
+            if let Some(line) = self.line.strip_suffix('\n') {
+                self.bytes.advance(read);
+
+                match serde_json::from_str(super::workaround_51132(line)) {
+                    Ok(JsonRow::Row(value)) => return Ok(Some(value)),
+                    Ok(JsonRow::Progress { .. }) => continue,
+                    Err(err) => return Err(Error::BadResponse(err.to_string())),
+                }
+            }
+
+            match self.raw.next().await? {
+                Some(chunk) => self.bytes.extend(chunk),
+                None => return Ok(None),
+            }
+        }
+    }
+}
